refactor: split write_finalization_layer
